5초마다 몬스터 재생성 + 몬스터 맵 밖으로 나가는거 해결

+++ Project/Object/GameObject/Brotato/Player/Player_Guns/Bro_Player_RailGun.h
@@ -15,7 +15,7 @@ public:
 	void SetActive(bool isActive);
 	bool IsActive() { return _isActive; }
 
-	UINT DMG = 40;
+	float DMG = 40.0f;
 private:
 	bool _isActive = true;
 	shared_ptr<Collider> _collider;
